[SQL] Break hiveOperators.scala into multiple files.
